Moved the friction folder into the newly created dflowfm folder and updated the references.
