+++ hydrolib/core/io/structure/models.py
@@ -14,7 +14,7 @@ from pydantic import Field
 from pydantic.class_validators import root_validator, validator
 
 from hydrolib.core.basemodel import DiskOnlyFileModel
-from hydrolib.core.io.friction.models import FrictionType
+from hydrolib.core.io.dflowfm.friction.models import FrictionType
 from hydrolib.core.io.dflowfm.ini.models import INIBasedModel, INIGeneral, INIModel
 from hydrolib.core.io.dflowfm.ini.util import (
     get_conditional_root_validator,
