+++ tests/io/test_friction.py
@@ -6,7 +6,7 @@ from typing import Any, Callable, List, Union
 import pytest
 from pydantic.error_wrappers import ValidationError
 
-from hydrolib.core.io.friction.models import (
+from hydrolib.core.io.dflowfm.friction.models import (
     FrictBranch,
     FrictGlobal,
     FrictionModel,
