+++ tests/io/test_structure.py
@@ -6,7 +6,7 @@ from typing import Any, List, Union
 import pytest
 from pydantic.error_wrappers import ValidationError
 
-from hydrolib.core.io.friction.models import FrictionType
+from hydrolib.core.io.dflowfm.friction.models import FrictionType
 from hydrolib.core.io.dflowfm.ini.parser import Parser, ParserConfig
 from hydrolib.core.io.structure.models import (
     Bridge,
